% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cic.R
\name{assert_same_length}
\alias{assert_same_length}
\title{Assert that N vectors are the same length}
\usage{
assert_same_length(...)
}
\arguments{
\item{...}{an arbitrary number of vectors}
}
\description{
Assert that N vectors are the same length
}
