% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cic.R
\name{build_cdfs}
\alias{build_cdfs}
\title{Gets a list of CDFs for every group}
\usage{
build_cdfs(i, t, x, ...)
}
\arguments{
\item{i}{index of individual}

\item{t}{index for time}

\item{x}{variable to construct the CDF of}

\item{g}{index for group}
}
\description{
Gets a list of CDFs for every group
}
