% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cic.R
\name{build_cic_data}
\alias{build_cic_data}
\title{Build CIC Data}
\usage{
build_cic_data(i, t, g, y, x, t_window = NA)
}
\arguments{
\item{i}{index of individual}

\item{t}{index for time}

\item{g}{index for group}

\item{y}{index for binary outcome}

\item{x}{variable to construct the CDF of}
}
\description{
Build CIC Data
}
