% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cic.R
\name{cic}
\alias{cic}
\title{Compute Changes in Changes model}
\usage{
cic(i, t, y, x, g = NULL, t_window = NULL, bootstrap = FALSE)
}
\arguments{
\item{i}{person index}

\item{t}{time index}

\item{y}{outcome index, y = 1 assumed to be what we proxy for}

\item{g}{optional group index}

\item{t_window}{time-window to restrict potential counterfactuals}
}
\description{
Compute Changes in Changes model
}
