% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cic.R
\name{cic_from_cicdata}
\alias{cic_from_cicdata}
\title{Run CIC given the build_cic_data output}
\usage{
cic_from_cicdata(cic_data, bootstrap)
}
\arguments{
\item{cic_data}{list output from cic_data}
}
\description{
Run CIC given the build_cic_data output
}
