% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cic.R
\name{cic_from_wecdfs}
\alias{cic_from_wecdfs}
\title{Takes in three vectors and computes the CIC model for the
initial one}
\usage{
cic_from_wecdfs(F_01, F_00, F_10)
}
\arguments{
\item{y_00}{vector of outcomes for group 0 at time 0}

\item{y_01}{vector of outcomes for group 0 at time 1}
}
\description{
Takes in three vectors and computes the CIC model for the
initial one
}
\details{
assume that group 1 is the "treated" group.
We are interested in the distribution of outcomes
(F^N_11) that would have occurred at time 1 for the
treated group had the intervention "treatment" not occurred
}
