% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counterfactuals.R
\name{combined_cf}
\alias{combined_cf}
\title{Generate combined counterfactual}
\usage{
combined_cf(cic)
}
\arguments{
\item{cic}{cic object retrned from function cic}
}
\description{
Generate combined counterfactual
}
\details{
If we think of the set of counterfactuals, {F_itsj}
where a cdf is indexed by a comparison group j and time j,
as being samples from the set of CDFs that would lead to
each class being obtained under the same distribution of
unobservables, then E(F_y1(x)) is the quantile of the expected
CDF that leads to outcome y1 and we can approximate it
with the empirical analogue.
}
