% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cic.R
\name{compute_cdfs}
\alias{compute_cdfs}
\title{takes in a list of data and computes empirical CDFs}
\usage{
compute_cdfs(l, ...)
}
\arguments{
\item{l}{list of numeric vectors to pass to wecdf}

\item{...}{other arguments to pass to wecdf}
}
\description{
takes in a list of data and computes empirical CDFs
}
