% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cic.R
\name{get_all_cfs}
\alias{get_all_cfs}
\title{Get all the counterfactuals for every person / time}
\usage{
get_all_cfs(i, t, g, y, t_window = NA)
}
\arguments{
\item{i}{index of individuals}

\item{t}{index of time}

\item{y}{index of outcome}

\item{t_window}{time window to filter on}

\item{group}{group to match on (for matching on observables)}
}
\description{
Get all the counterfactuals for every person / time
}
