% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cic.R
\name{get_cdf}
\alias{get_cdf}
\title{Helper function, gets a cdf from a list given an index}
\usage{
get_cdf(cdflist, i, t)
}
\arguments{
\item{cdflist}{list of cdfs from construct_cdfs}

\item{i}{person index}

\item{t}{time index}

\item{g}{group index}
}
\description{
Helper function, gets a cdf from a list given an index
}
