% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cic.R
\name{get_itg_y1}
\alias{get_itg_y1}
\title{Get the observations for i and t such that y = 1}
\usage{
get_itg_y1(i, t, g, y)
}
\arguments{
\item{i}{person index}

\item{t}{time index}
}
\description{
Get the observations for i and t such that y = 1
}
