% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cic.R
\name{get_x}
\alias{get_x}
\title{Get the training data for an empirical CDF}
\usage{
get_x(e)
}
\arguments{
\item{e}{empirical distribution function to get
stuff from}
}
\description{
Get the training data for an empirical CDF
}
