% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cic.R
\name{id_cfs}
\alias{id_cfs}
\alias{js_00}
\alias{js_01}
\alias{js_10}
\alias{js_11}
\title{Identify counterfactuals from data}
\usage{
id_cfs(i, t_i, g_i, j, t, g, y, t_window = NA)

js_00(i, t_i, g_i, j, t, y, g)

js_01(i, t_i, g_i, j, t, y, g)

js_10(i, t_i, g_i, j, t, y, g)

js_11(i, t_i, g_i, j, t, y, g)
}
\arguments{
\item{i}{a scalar / single identifier for which group to consider}

\item{t_i}{time for outcome y = 1}

\item{g_i}{group of the individual i for matching on observables}

\item{j}{index of identifiers for groups}

\item{t}{index of years}

\item{y}{outcomes for each group x year}
}
\description{
Identify counterfactuals from data
}
\details{
Note that y_i = 1 by convention, this function assumes
that indexing.
}
