% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cic.R
\name{print.cic}
\alias{print.cic}
\title{Print function for cic}
\usage{
\method{print}{cic}(x, ...)
}
\arguments{
\item{x}{fitted cic model}

\item{...}{other arguments to print}
}
\description{
Print function for cic
}
