% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_ecdf.R
\name{quantile.wecdf}
\alias{quantile.wecdf}
\title{Method for the quantile function for weighted empriical CDFs}
\usage{
\method{quantile}{wecdf}(x, probs, ...)
}
\arguments{
\item{x}{empirical cdf function of class weighted_cdf}

\item{probs}{vector of probabilities to evaluate quantile function for}

\item{...}{other things to pass to weighted quantile}

\item{na.rm}{whether to remove NAs}
}
\description{
Method for the quantile function for weighted empriical CDFs
}
