% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cic.R
\name{randweight_cdf}
\alias{randweight_cdf}
\title{Reweight CDFs with random exponentials}
\usage{
randweight_cdf(cdf)
}
\arguments{
\item{cdf}{cumulative distribution to randomly weight}
}
\description{
Reweight CDFs with random exponentials
}
\details{
Used for bootstrapping but nothing else
}
