% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cic.R
\name{rdirichlet}
\alias{rdirichlet}
\title{Simulate from dirichlet distribution}
\usage{
rdirichlet(n, alpha = 1)
}
\arguments{
\item{n}{number of simulations}

\item{alpha}{alpha to be passed (assumed constant)}
}
\description{
Simulate from dirichlet distribution
}
