% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cic.R
\name{set_parallel}
\alias{set_parallel}
\title{Set parallel settings}
\usage{
set_parallel(n_workers, ...)
}
\arguments{
\item{n_workers}{number of workers}

\item{...}{other arguments passed to future::plan}
}
\description{
Set parallel settings
}
