% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_ecdf.R
\name{wecdf}
\alias{wecdf}
\title{Weighted empirical cdf function}
\usage{
wecdf(x, weights = NULL, na.rm = FALSE, normalize_weights = TRUE)
}
\arguments{
\item{x}{vector to calculate the weighted empirical CDF for}

\item{weights}{vector of weights, must not contain NAs}

\item{na.rm}{whether to remove NAs from x and associated weight vector}

\item{normalize_weights}{whether to normalize weights to sum to 1}
}
\description{
Weighted empirical cdf function
}
\details{
This is _heavily_ based on the code from the
ggdist package. Thanks to Matthew Kay and Brenton Wiernik
for writing a lovely package. I would have imported it
but I didn't want a whole dependency and wanted some additional
checks.
}
